# backend/app/models/enums.py
from __future__ import annotations

from enum import StrEnum


class UploadType(StrEnum):
    PDF = "PDF"
    TEXT = "TEXT"


class DocumentStatus(StrEnum):
    UPLOADED = "UPLOADED"
    TEXT_EXTRACTING = "TEXT_EXTRACTING"
    TEXT_EXTRACTED = "TEXT_EXTRACTED"
    REFERENCES_EXTRACTING = "REFERENCES_EXTRACTING"
    REFERENCES_EXTRACTED = "REFERENCES_EXTRACTED"
    DOI_VERIFYING = "DOI_VERIFYING"
    DOI_VERIFIED = "DOI_VERIFIED"
    CLAIMS_EXTRACTING = "CLAIMS_EXTRACTING"
    CLAIMS_EXTRACTED = "CLAIMS_EXTRACTED"
    EVIDENCE_PREPARING = "EVIDENCE_PREPARING"
    EVIDENCE_READY = "EVIDENCE_READY"
    VERIFYING = "VERIFYING"
    VERIFIED = "VERIFIED"
    REPORT_GENERATED = "REPORT_GENERATED"
    FAILED = "FAILED"
    PARTIAL_FAILED = "PARTIAL_FAILED"


class PipelineStatus(StrEnum):
    QUEUED = "QUEUED"
    RUNNING = "RUNNING"
    SUCCEEDED = "SUCCEEDED"
    FAILED = "FAILED"
    PARTIAL_FAILED = "PARTIAL_FAILED"
    CANCELLED = "CANCELLED"


class PipelineStepStatus(StrEnum):
    PENDING = "PENDING"
    QUEUED = "QUEUED"
    RUNNING = "RUNNING"
    SUCCEEDED = "SUCCEEDED"
    FAILED = "FAILED"
    SKIPPED = "SKIPPED"
    PARTIAL_FAILED = "PARTIAL_FAILED"


class DoiStatus(StrEnum):
    FOUND = "FOUND"
    MISSING = "MISSING"
    MALFORMED = "MALFORMED"
    VALID = "VALID"
    INVALID = "INVALID"
    LOOKUP_FAILED = "LOOKUP_FAILED"


class MetadataStatus(StrEnum):
    NOT_LOOKED_UP = "NOT_LOOKED_UP"
    LOOKUP_SUCCEEDED = "LOOKUP_SUCCEEDED"
    LOOKUP_FAILED = "LOOKUP_FAILED"
    METADATA_UNAVAILABLE = "METADATA_UNAVAILABLE"


class ClaimType(StrEnum):
    EMPIRICAL = "EMPIRICAL"
    THEORETICAL = "THEORETICAL"
    METHODOLOGICAL = "METHODOLOGICAL"
    CAUSAL = "CAUSAL"
    COMPARATIVE = "COMPARATIVE"
    DEFINITION = "DEFINITION"
    BACKGROUND = "BACKGROUND"
    UNKNOWN = "UNKNOWN"


class MappingStatus(StrEnum):
    MAPPED = "MAPPED"
    UNCERTAIN = "UNCERTAIN"
    NO_MATCH = "NO_MATCH"
    MULTIPLE_MATCHES = "MULTIPLE_MATCHES"
    NEEDS_HUMAN_REVIEW = "NEEDS_HUMAN_REVIEW"


class EvidenceAvailability(StrEnum):
    METADATA_ONLY = "METADATA_ONLY"
    ABSTRACT_AVAILABLE = "ABSTRACT_AVAILABLE"
    PREPRINT_AVAILABLE = "PREPRINT_AVAILABLE"
    FULL_TEXT_AVAILABLE = "FULL_TEXT_AVAILABLE"
    SOURCE_UNAVAILABLE = "SOURCE_UNAVAILABLE"


class RetrievalStatus(StrEnum):
    SUCCEEDED = "SUCCEEDED"
    FAILED = "FAILED"
    PARTIAL_FAILED = "PARTIAL_FAILED"
    NO_RELEVANT_EVIDENCE_FOUND = "NO_RELEVANT_EVIDENCE_FOUND"
    SERVICE_UNAVAILABLE = "SERVICE_UNAVAILABLE"
    TIMEOUT = "TIMEOUT"


class SupportStatus(StrEnum):
    SUPPORTED = "SUPPORTED"
    PARTIALLY_SUPPORTED = "PARTIALLY_SUPPORTED"
    NOT_SUPPORTED = "NOT_SUPPORTED"
    INSUFFICIENT_EVIDENCE = "INSUFFICIENT_EVIDENCE"
    NEEDS_HUMAN_REVIEW = "NEEDS_HUMAN_REVIEW"


class CacheSource(StrEnum):
    NEW_VERIFICATION = "NEW_VERIFICATION"
    EXACT_CACHE = "EXACT_CACHE"
    SEMANTIC_CACHE = "SEMANTIC_CACHE"
    HUMAN_CORRECTED = "HUMAN_CORRECTED"


class SafetyRiskLevel(StrEnum):
    LOW = "LOW"
    MEDIUM = "MEDIUM"
    HIGH = "HIGH"
    UNKNOWN = "UNKNOWN"